# docs/source/_static/trustpoint-openapi-swagger.yaml
openapi: 3.0.3
info:
  title: Trustpoint APIs
  version: v0.0.5
  description: API documentation for Trustpoint project
paths:
...
  /api/certificates/{id}/:
    get:
      operationId: api_certificates_retrieve
      description: |-
        ViewSet for managing Certificate instances.

        Supports standard CRUD operations such as list, retrieve,
        create, update, and delete.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this certificate model.
        required: true
      tags:
      - Certificate
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
          description: ''
    put:
      operationId: api_certificates_update
      description: |-
        ViewSet for managing Certificate instances.

        Supports standard CRUD operations such as list, retrieve,
        create, update, and delete.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this certificate model.
        required: true
      tags:
      - Certificate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Certificate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Certificate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Certificate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
          description: ''
    patch:
      operationId: api_certificates_partial_update
      description: |-
        ViewSet for managing Certificate instances.

        Supports standard CRUD operations such as list, retrieve,
        create, update, and delete.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this certificate model.
        required: true
      tags:
      - Certificate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCertificate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCertificate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCertificate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
          description: ''
    delete:
      operationId: api_certificates_destroy
      description: |-
        ViewSet for managing Certificate instances.

        Supports standard CRUD operations such as list, retrieve,
        create, update, and delete.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this certificate model.
        required: true
      tags:
      - Certificate
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
...
    Certificate:
      type: object
      description: |-
        Serializer for Certificate instances.

        Handles conversion between Certificate model objects and JSON representations.
      properties:
        id:
          type: integer
          readOnly: true
        common_name:
          type: string
          maxLength: 256
      required:
      - id
...
    PatchedCertificate:
      type: object
      description: |-
        Serializer for Certificate instances.

        Handles conversion between Certificate model objects and JSON representations.
      properties:
        id:
          type: integer
          readOnly: true
        common_name:
          type: string
          maxLength: 256
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT